SubjectResponse에 Subject ID 추가

+++ src/main/java/com/project/mentoridge/modules/lecture/repository/dto/LectureMentorQueryDto.java
@@ -6,10 +6,10 @@ import lombok.Data;
 public class LectureMentorQueryDto {
 
     private Long mentorId;
-    private Long lectureCount;   // 총 강의 수
-    private Long reviewCount;    // 리뷰 개수
+    private int lectureCount;   // 총 강의 수
+    private int reviewCount;    // 리뷰 개수
 
-    public LectureMentorQueryDto(Long mentorId, Long lectureCount, Long reviewCount) {
+    public LectureMentorQueryDto(Long mentorId, int lectureCount, int reviewCount) {
         this.mentorId = mentorId;
         this.lectureCount = lectureCount;
         this.reviewCount = reviewCount;
