+++ src/main/java/com/project/mentoridge/modules/lecture/repository/dto/LectureReviewQueryDto.java
@@ -6,11 +6,11 @@ import lombok.Data;
 public class LectureReviewQueryDto {
 
     private Long lectureId;
-    private Long reviewCount;
+    private int reviewCount;
     // private Integer reviewCount;    // 리뷰 총 개수
     private double scoreAverage;    // 강의 평점
 
-    public LectureReviewQueryDto(Long lectureId, Long reviewCount, double scoreAverage) {
+    public LectureReviewQueryDto(Long lectureId, int reviewCount, double scoreAverage) {
         this.lectureId = lectureId;
         this.reviewCount = reviewCount;
         this.scoreAverage = scoreAverage;
